Split benchmarks out of main source
